CHROMIUM: cpuquiet: Extract non-arch specific code from cpuquiet-tegra into cpuquiet-common.c.
BUG=chrome-os-partner:40516
TEST=Build and boot.

Reviewed-on: https://chromium-review.googlesource.com/218060
Reviewed-on: https://chromium-review.googlesource.com/272859
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
