+++ drivers/cpuquiet/Makefile
@@ -2,4 +2,4 @@ GCOV_PROFILE := y
 
 obj-$(CONFIG_CPU_QUIET) += driver.o governor.o governors/
 obj-$(CONFIG_CPU_QUIET_STATS) += sysfs.o attribute.o
-obj-$(CONFIG_TEGRA_CPUQUIET) += cpuquiet-tegra.o
+obj-$(CONFIG_TEGRA_CPUQUIET) += cpuquiet-common.o cpuquiet-tegra.o
